fix warning

+++ src/media_tools/media_import.c
@@ -5552,7 +5552,7 @@ restart_import:
 		}
 
 		if (min_cts_offset > 0) {
-			gf_isom_shift_cts_offset(import->dest, track, min_cts_offset);
+			gf_isom_shift_cts_offset(import->dest, track, (s32)min_cts_offset);
 		}
 		/*and repack table*/
 		gf_isom_set_cts_packing(import->dest, track, GF_FALSE);
